backend/requirements.txt
akshare==1.18.21
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl-cffi==0.14.0
decorator==5.2.1
distro==1.9.0
duckduckgo-search==8.1.1
et-xmlfile==2.0.0
fastapi==0.128.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpath==0.82.2
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-openai==1.1.7
langsmith==0.6.7
lxml==6.0.2
macholib==1.16.4
mini-racer==0.14.1
numpy==2.4.2
openai==2.16.0
openpyxl==3.1.5
orjson==3.11.6
packaging==25.0
pandas==3.0.0
primp==0.15.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.0
uvicorn==0.40.0
webencodings==0.5.1
xlrd==2.0.2
xxhash==3.6.0
zstandard==0.25.0
cryptography==46.0.4
pymysql==1.1.1
DBUtils==3.1.0
